Delete account command with masstransit

+++ src/Ubik.Accounting.Api/Features/Accounts/Commands/DeleteAccount.cs
@@ -2,43 +2,40 @@
 using MediatR;
 using System.ComponentModel.DataAnnotations;
 using Ubik.Accounting.Api.Features.Accounts.Exceptions;
+using Ubik.Accounting.Api.Features.Accounts.Mappers;
+using Ubik.Accounting.Contracts.Accounts.Commands;
 using Ubik.Accounting.Contracts.Accounts.Events;
+using Ubik.Accounting.Contracts.Accounts.Results;
 
 namespace Ubik.Accounting.Api.Features.Accounts.Commands;
-public class DeleteAccount
+
+public class DeleteAccountConsumer : IConsumer<DeleteAccountCommand>
 {
-    //Input
-    public record DeleteAccountCommand : IRequest<bool>
+    private readonly IServiceManager _serviceManager;
+    private readonly IPublishEndpoint _publishEndpoint;
+
+    public DeleteAccountConsumer(IServiceManager serviceManager, IPublishEndpoint publishEndpoint)
     {
-        [Required]
-        public Guid Id { get; set; }
+        _serviceManager = serviceManager;
+        _publishEndpoint = publishEndpoint;
     }
 
-
-    public class DeleteAccountHandler : IRequestHandler<DeleteAccountCommand,bool>
+    public async Task Consume(ConsumeContext<DeleteAccountCommand> context)
     {
-        private readonly IServiceManager _serviceManager;
-        private readonly IPublishEndpoint _publishEndpoint;
+        //Check if the account is found
+        var account = await _serviceManager.AccountService.GetAsync(context.Message.Id);
 
-        public DeleteAccountHandler(IServiceManager serviceManager, IPublishEndpoint publishEndpoint)
+        if (account == null)
         {
-            _serviceManager = serviceManager;
-            _publishEndpoint = publishEndpoint;
+            await context.RespondAsync(new AccountNotFoundException(context.Message.Id));
+            return;
         }
 
-        public async Task<bool> Handle(DeleteAccountCommand request, CancellationToken cancellationToken)
-        {
-            //Check if the account is found
-            var account = await _serviceManager.AccountService.GetAsync(request.Id)
-                            ?? throw new AccountNotFoundException(request.Id);
+        await _serviceManager.AccountService.ExecuteDeleteAsync(account.Id);
+        await _publishEndpoint.Publish(new AccountDeleted { Id = account.Id }, CancellationToken.None);
+        await _serviceManager.SaveAsync();
 
-
-            await _serviceManager.AccountService.ExecuteDeleteAsync(account.Id);
-            await _publishEndpoint.Publish(new AccountDeleted { Id= account.Id }, CancellationToken.None);
-            await _serviceManager.SaveAsync();
-
-            return true;
-        }
+        await context.RespondAsync <DeleteAccountResult>(new {Deleted = true});
     }
 }
